+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Aurora PostgreSQL CloudWatch with slow query auditing example
 - Google Cloud SQL for MySQL module
 - Google Cloud SQL for PostgreSQL module
+- Google Cloud SQL for SQL Server module
 
 ### Bug Fixes
 - Modified Server Host Name of AWS RDS MS SQL SERVER Dsfhub assets